// src/Interfaces/IPerson.ts
export default interface IPerson {
  CPF: string;
  address: string;
  city: string;
  coutry: string;
  firstName: string;
  id: number;
  image: string;
  jobArea: string;
  jobDescriptor: string;
  jobTitle: string;
  jobType: string;
  lastName: string;
  latitude: number;
  longitude: number;
  phone: string;
  state: string;
  zipCode: string;
}